Added Projectile class
Projectiles will be used by the Soldier enemy to shoot at the player. getX() and getY() methods have been added to the Player class to allow that. At the moment a projectile is shot from the middle of the canvas to the player everytime a click is detected, should be changed so the draw() method in Soldier creates a Projectile aimed at the current location of the player every few seconds.